+++ addon/components/sweet-alert.js
@@ -1,6 +1,6 @@
-import Component from '@ember/component';
+import Component from '@glimmer/component';
 import { inject as service } from '@ember/service';
-import { get, set } from '@ember/object';
+import { action } from '@ember/object';
 import { A } from '@ember/array';
 
 const CONFIGURATION = [
@@ -57,63 +57,56 @@ const CONFIGURATION = [
   'progressSteps',
   'currentProgressStep',
   'progressStepsDistance',
-  'onBeforeOpen',
-  'onOpen',
-  'onClose'
 ];
 
-const SweetAlertComponent = Component.extend({
-  swal: service(),
+const EVENTS = [
+  'onBeforeOpen',
+  'onOpen',
+  'onClose',
+];
 
-  show: true,
-  onConfirm: () => {},
-  onCancel: () => {},
+export default class SweetAlertComponent extends Component {
+  @service swal;
 
-  didInsertElement() {
-    this._super(...arguments);
-    this._displaySweetAlert();
-  },
-
-  didUpdateAttrs() {
-    this._super(...arguments);
-    this._displaySweetAlert();
-  },
-
-  _displaySweetAlert() {
-    if (get(this, 'show')) {
-      let props = this._getValues();
-
-      get(this, 'swal').open(props).then((result) => {
-        if (result.value) {
-          get(this, 'onConfirm')(result);
-        } else {
-          get(this, 'onCancel')(result);
-        }
-
-        if (!get(this, 'isDestroyed')) {
-          set(this, 'show', false);
-        }
-      });
+  get isOpen() {
+    if (undefined === this.args.show) {
+      return true;
     }
-  },
 
-  _getValues() {
+    return this.args.show;
+  }
+
+  @action async fire() {
+    let result = await this.swal.open(this._values());
+
+    if (result.value) {
+      this._call('onConfirm', result);
+    } else {
+      this._call('onCancel', result);
+    }
+  }
+
+  _call(method, ...args) {
+    if (!this.isDestroying && this.args[method]) {
+      this.args[method](...args);
+    }
+  }
+
+  _values() {
     let props = {};
 
     A(CONFIGURATION).forEach(key => {
-      let value = get(this, key);
+      let value = this.args[key];
 
       if (undefined !== value) {
         props[key] = value;
       }
     });
 
+    A(EVENTS).forEach(key => {
+      props[key] = () => this._call(key, ...arguments);
+    });
+
     return props;
   }
-});
-
-SweetAlertComponent.reopenClass({
-  positionalParams: ['title', 'text', 'type']
-});
-
-export default SweetAlertComponent;
+}